% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocManager-pkg.R
\docType{package}
\name{BiocManager-pkg}
\alias{BiocManager-pkg}
\alias{_PACKAGE}
\alias{BiocManager}
\alias{BiocManager-pkg-package}
\title{Install or update Bioconductor, CRAN, or GitHub packages}
\description{
This package provides tools for managing \emph{Bioconductor} and other
packages in a manner consistent with \emph{Bioconductor}'s package
versioning and release system.
}
\details{
Main functions are as follows; additional help is
available for each function, e.g., \code{?BiocManager::version}.

\describe{

\item{\code{BiocManager::install()}}{Install or update packages from
\emph{Bioconductor}, CRAN, and GitHub.}

\item{\code{BiocManager::version()}}{Report the version of
\emph{Bioconductor} in use.}

\item{\code{BiocManager::available()}}{Return a \code{character()} vector
of package names available (at \code{BiocManager::repositories()})
for installation.}

\item{\code{BiocManager::valid()}}{Determine whether installed packages
are from the same version of \emph{Bioconductor}.}

\item{\code{BiocManager::repositories()}}{\emph{Bioconductor} and other
repository URLs to discover packages for installation.}

}
}
\author{
\strong{Maintainer}: Bioconductor Package Maintainer \email{maintainer@bioconductor.org} [contributor]

Authors:
\itemize{
  \item Martin Morgan \email{martin.morgan@roswellpark.org}
}

Other contributors:
\itemize{
  \item Marcel Ramos \email{marcel.ramos@roswellpark.org} [contributor]
}

}
