% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocUpdatePackages.R
\name{biocUpdatePackages}
\alias{biocUpdatePackages}
\title{Update previously installed Bioconductor or CRAN packages and their
dependencies.}
\usage{
biocUpdatePackages(pkgs, dependencies = NA, repos = biocinstallRepos(), ...)
}
\arguments{
\item{pkgs}{\code{character()} of package names to install or update.}

\item{dependencies}{\code{character()} describing out-of-date dependencies
that are also updated. Defaults to \code{c("Depends", "Imports",
"LinkingTo")} but can be a subset of \code{c("Depends", "Imports",
"LinkingTo", "Suggests", "Enhances")}.}

\item{repos}{\code{character()} of named repositories in which to look for
package updates, in the style of \code{biocinstallRepos()}.}

\item{\dots}{Additional arguments, passed to \code{\link{update.packages}}.
For example, \code{ask=FALSE} to avoid prompts to update individual
packages.}
}
\value{
\sQuote{NULL}, invisibly.
}
\description{
Update previously installed Bioconductor and CRAN packages and their
dependencies. Use \code{\link{biocLite}} to install new packages or to
update all out-of-date packages. Upgrading to a new Bioconductor release
requires additional steps; see \url{https://bioconductor.org/install}.
}
\examples{

\dontrun{
biocUpdatePackages("GenomicRanges", ask=FALSE)
}

}
\seealso{
\code{\link{biocLite}}
}
\author{
Martin Morgan \url{mtmorgan@fhcrc.org}
}
\keyword{environment}
