% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocVersion.R
\name{biocVersion}
\alias{biocVersion}
\title{Bioconductor version}
\usage{
biocVersion()
}
\value{
\code{package_version} representing the Bioconductor version in use.
}
\description{
This function reports the Bioconductor version in use, as determined by the
BiocInstaller package.
}
\examples{

biocVersion()

}
\seealso{
\code{\link{install}} Installs/updates Bioconductor/CRAN packages.
}
\keyword{environment}
