% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installRepos.R
\name{installRepos}
\alias{installRepos}
\title{Display current Bioconductor and CRAN repositories.}
\usage{
installRepos(siteRepos = character(), version = biocVersion())
}
\arguments{
\item{siteRepos}{(Optional) \code{character(1)} representing an additional
repository (e.g., a URL to an organization's internally maintained
repository) in which to look for packages to install. This repository will
be prepended to the default repositories returned by the function.}

\item{version}{(Optional) \code{character(1)} or \code{package_version}
indicating the Bioconductor version (e.g., \dQuote{3.1}) for which
repositories are required.}
}
\value{
Named \code{character()} of repositories.
}
\description{
Displays the URLs of the repositories used by \code{\link{install}} to
install Bioconductor and CRAN packages.
}
\examples{

installRepos()

## Choose mirrors
\dontrun{
  chooseCRANmirror()
  chooseBioCmirror()
}

}
\seealso{
\code{\link{install}} Installs/updates Bioconductor/CRAN packages.

\code{\link{install.packages}} installs the packages themselves.

\code{\link{chooseBioCmirror}} lets you choose from a list of all public
  Bioconductor mirror URLs.

\code{\link{chooseCRANmirror}} lets you choose from a list of all public
  CRAN mirror URLs.
}
\keyword{environment}
