% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot}
\alias{snapshot}
\title{Check whether your Bioconductor version has a snapshot}
\usage{
snapshot(version = BiocManager::version())
}
\arguments{
\item{version}{character(1) A Bioconductor version number, as given by
\code{BiocManager::version()}.}
}
\value{
A \code{data.frame} of mapped R and Bioconductor versions and their
respective snapshot links where available
}
\description{
This function provides the version map located in
\url{https://bioconductor.org/config.yaml} for a particular Bioconductor
version. For out-of-date Bioconductor versions, it is recommended that
the user use an appropriate snapshot of the packages in CRAN history that
coincides with the last-built date of that Bioconductor version. \code{snapshot}
gives the user a \code{data.frame} of MRAN snapshots and RSPM repositories
available for a particular Bioconductor version. Note that there can only
be one CRAN repository set at a time. Recent versions of Bioconductor
do not need to use a snapshot and the default CRAN repository is
recommended.
}
\details{
To make persistent changes across sessions to \code{options("repos")},
it is recommended that users enter the adequate CRAN snapshot in
their \code{.Rprofile} file as given by \code{snapshot}. This file is usually
located under the user's home directory, see \code{normalizePath("~")}. The
\code{options()} function call should look like the following:
\preformatted{
        options(
            repos = c(CRAN = "https://mran.microsoft.com/snapshot/date")
        )
    }
}
\section{MRAN}{

The Microsoft R Archive Network provides a historical repository of CRAN
packages by date. The date in the snapshot link is formatted as:
\preformatted{
        https://mran.microsoft.com/snapshot/YYYY-MM-DD
    }
}

\examples{

snapshot("3.2")

}
