% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useDevel.R
\name{useRelease}
\alias{useRelease}
\title{Use the release version of Bioconductor}
\usage{
useRelease()
}
\description{
This function will modify the repository pointer to the release version
if you are using devel. It will then prompt a downgrade of all packages
in the \code{.libPaths()[1]}.
}
