% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\title{Validate installed package versions against correct versions.}
\usage{
valid(pkgs = installed.packages(lib.loc, priority = priority),
  lib.loc = NULL, priority = "NA", type = getOption("pkgType"),
  filters = NULL, silent = FALSE, ..., fix = FALSE)
}
\arguments{
\item{pkgs}{A character() vector of package names for checking, or
a matrix as returned by \code{\link{installed.packages}}.}

\item{lib.loc}{A character() vector of library location(s) of
packages to be validated; see \code{\link{installed.packages}}.}

\item{priority}{character(1) Check validity of all, \dQuote{base},
or \dQuote{recommended} packages; see
\code{\link{installed.packages}}.}

\item{type}{character(1) The type of available package (e.g.,
binary, source) to check validity against; see
\code{\link{available.packages}}.}

\item{filters}{character(1) Filter available packages to check
validity against; see \code{\link{available.packages}}.}

\item{silent}{logical(1) Report how packages are invalid
(\code{silent=FALSE}, default) and abort execution, or return a
logical(1) (\code{silent=TRUE}) indicating the overall validity
of installed packages.}

\item{\dots}{Additional arguments, passed to \code{\link{install}}
when \code{fix=TRUE}.}

\item{fix}{When \code{TRUE}, invoke \code{install} to reinstall
(update or downgrade, as appropriate) invalid packages.}
}
\value{
\code{logical(1)} indicating overall validity of installed
    packages.
}
\description{
Check that installed packages are consistent (neither out-of-date
nor too new) with the version of R and Bioconductor in use, using
\code{install} for validation.
}
\details{
This function compares the version of installed packages to the
version of packages associated with the version of R and
Bioconductor appropriate for the BiocManager package currently in
use.

Packages are reported as \sQuote{out-of-date} if a more recent
version is available at the repositories specified by
\code{repositories()}.  Usually, \code{install()} is sufficient to
update packages to their most recent version.

Packages are reported as \sQuote{too new} if the installed version
is more recent than the most recent available in the
\code{repositories()} repositories. It is possible to down-grade by
re-installing a too new package \dQuote{PkgA} with
\code{install("PkgA")}. It is important for the user to understand
how their installation became too new, and to avoid this in the
future.
}
\examples{
try(valid())
}
\seealso{
\code{\link{install}} to update installed packages.
}
\author{
Martin Morgan \email{martin.morgan@roswellpark.org}
}
\keyword{environment}
