% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Version of Bioconductor currently in use.}
\usage{
version()
}
\value{
A two-digit version, e.g., \code{3.8}, of class
\code{package_version} describing the version of \emph{Bioconductor} in
use.
}
\description{
This is the version of \emph{Bioconductor} appropropriate for this
version of R or, if the user has requested, the 'devel' version of
\emph{Bioconductor}.
}
\examples{
version()

}
