% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioconductor-pkg.R
\docType{package}
\name{BiocManager-pkg}
\alias{BiocManager-pkg}
\alias{_PACKAGE}
\alias{BiocManager}
\alias{BiocManager-pkg-package}
\title{Install or update Bioconductor, CRAN, or GitHub packages}
\description{
Install or update Bioconductor, CRAN, or GitHub packages
}
\author{
\strong{Maintainer}: Bioconductor Package Maintainer \email{maintainer@bioconductor.org} [contributor]

Authors:
\itemize{
  \item Martin Morgan \email{martin.morgan@roswellpark.org}
}

Other contributors:
\itemize{
  \item Marcel Ramos \email{marcel.ramos@roswellpark.org} [contributor]
}

}
