% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Version of Bioconductor currently installed}
\usage{
version()
}
\value{
A two-digit version, e.g., \code{'3.7'}, of class
\code{package_version} describing the version of \emph{Bioconductor} in
use. A version of \code{'0.0'} indicates that information on
\emph{Bioconductor} version is not available (i.e., the BiocVersion
package is not installed).
}
\description{
The \code{install()} function arranges to install the BiocVersion
package. The BiocVersion package has a version number that
corresponds to the version of \emph{Bioconductor} in use. \code{version()}
reports on this version. If BiocVersion is not intalled,
\code{version()} returns a sentinel (version '0.0').
}
\examples{
version()

}
