% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\alias{print.version_sentinel}
\title{Version of Bioconductor currently in use.}
\usage{
version()

\method{print}{version_sentinel}(x, ...)
}
\arguments{
\item{x}{An \code{unknown_version} instance used to represent the
situation when the version of Bioconductor in use cannot be
determined.}

\item{...}{Additional arguments, ignored.}
}
\value{
A two-digit version, e.g., \code{3.8}, of class
\code{package_version} describing the version of \emph{Bioconductor} in
use.
}
\description{
\code{version()} reports the version of \emph{Bioconductor} appropropriate
for this version of R, or the version of \emph{Bioconductor} requested
by the user.
}
\details{
\code{version()} (and all functions requiring version information) fails
when version cannot be validated e.g., because internet access is
not available.
}
\examples{
BiocManager::version()

}
